TestArrayGroup: Added tests for arrays and DynamicData
The thorn provides a parameter file for these tests. Arrays
(distrib=const) are written in one function and read in another function.
The DynamicData test calls the DynamicData function for a grid function,
array, and scalar and compares the data to the correct values. These tests
also serve as a basic test of the validity mechanisms. The initialize
function writes all three types of variables everywhere. Then, the DynamicData
function reads all of them everywhere.