CarpetX: Fixed incorrect handling of interpolation data
The new interface was incorrectly handling the data from the
table handles, as the test was providing the data to the table
in a non-standard way. The test was fixed, and the code now handles
the expected inputs properly.
This commit also overloads the InterpGridArrays function. The code
was present, but the line to overload it in Cactus was not included.